requirements.txt
flask
flask-cors
python-dotenv
requests
google-generativeai